pyproject.toml
[project]
name = "ai-crm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
]
